+++ NEWS.md
@@ -48,6 +48,8 @@
   and `predict()`. Now much faster for large models. (version `2.8.0.9009`)
 * Fixed linearisation issue when using only the `*_latent` form of a component.
   (version `2.8.0.9015`)
+* Workaround for equivalent but textually different CRS/WKT information in
+  `bru_fill_missing()`. (version `2.9.0.9016`)
 
 ## Deprecation of old functions